// applications/webapp/src/app/api/generated/model/create-pipeline-run-response.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ItemResult } from './item-result';


export interface CreatePipelineRunResponse { 
    runId?: string;
    items?: Array<ItemResult>;
}