// applications/webapp/src/app/api/generated/model/youtube-channel-video-summary.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface YoutubeChannelVideoSummary { 
    id?: string;
    youtubeVideoId?: string;
    title?: string;
    publishedAt?: string;
    watchUrl?: string;
    ingested?: boolean;
}